Fix hanging notes on exit when using Portmidi (fixes #232) (#233)
* Add Portmidi note killing functionality

When using the PortMidi player, e.g. in an ALSA MIDI setup, closing the
program often leaves some hanging notes, since they aren't stopped by
the game when it exits. This fixes it by sending a MIDI Stop All Sound
message to all channels in the shutdown callback, right before actually
shutting PortMidi itself down.

See issue #232.

* Fix pm_killnotes default argument error

C does not support default arguments. I defined a 'when' argument that
defaulted to 0. Now, it should be specified manually. Just typing a
single zero is not all that different from an empty argument list, you
know!

* Clarify pm_killnotes 'when' default argument

The code automatically sets 'when' to Pt_Time() if 0 is passed. I have
extended a comment a tad, to help make this behaviour clearer, mostly
for maintainability purposes.

* Add forward declaration for writeevent

Shuffling the functions is probably a better idea, but there's
something about having pm_killnotes 'grouped' with other pm_*
functions. Either way, this is good enough to avert implicit
declaration related errors emitted by the compiler.

* Solve hanging notes by sending All Notes Off

Previously I attempted to kill potentially hanging notes by sending an
All Sound Off MIDI event. This did not seem to work. Plus, even if it
did, it would probably result in notes being sharply (and unnaturally)
cut off, instead of a gentle fade out.

This attempts to rectify those issues by sending an All Notes Off
message instead.

* Use pm_stop instead of pm_killnotes

pm_stop already does the functionality that is was expected from
pm_killnotes, the latter which turned out to be a redundant function.
It was removed accordingly, and pm_stop is being used in its place in
pm_shutdown.

* Forward declare pm_stop so pm_shutdown can use it

* Set when to 0 instead of Pt_Now() in pm_stop()

* Made Pt_Sleep call non-platform-dependent

So it turns out this call is useful not only in Win32. It prevents
hanging notes in Portmidi under Linux, at least when using ALSA.
As for updating that comment above the Pt_Sleep call, I will
leave it to Fabian to decide what to do with it.

* Revert change to definition of when in pm_stop