Restrict Alt+Tab workaround to SDL versions <2.0.18 (#447)
The latest SDL2 version (2.0.18) allows Alt+Tab to work as intended. Moreover, using the SDL_HINT_VIDEO_MINIMIZE_ON_FOCUS_LOSS workaround with it loses the mouse when Alt-tabing away from a fullscreen prb+ session. Thus we restrict the workaround to the two earlier versions, the way it's currently done in Choco.