Completely initialize blocklinks in P_UnArchiveBlockLinks
When loading save game / restoring keyframe on maps with lot's of
monsters and friction floor, it can happen that Doom is blocked on
indefinite loop.

Problem is loop in 'blocklinks' linked list. It is cased by fact, that
'blocklinks' is initialized 2 times (in 'P_SetThingPosition' and in
'P_TrueUnArchiveThinkers', both called from 'P_TrueUnArchiveThinkers')
and these two sources can be unsynchronized (in case of friction floors,
'P_Move' updates coordinates "after" calling 'P_TryMove', so it is
possible that coordinates does not correspond to position in
'blocklist').