CMake: Set _FILE_OFFSET_BITS=64 on Linux
By default without _GNU_SOURCE, off64_t is not defined. This breaks
MUSIC/dumbplayer.c workaround for making off_t 64-bit. Even 64-bit Linux
is affected.

Define _FILE_OFFSET_BITS=64 explicitly so that off_t is defined as a
64-bit type. This also skips the workaround in dumbplayer.c source file.

The `UNIX AND NOT APPLE` is assumed to mean `Linux`, though it might
trigger on non-Linux systems as well.