gl_preprocess: handle more self-referencing sectors
When a sector can't be tessellated as a whole because it's
not closed, the GL renderer falls back on subsector tesselation.
If the sector in question contains self-referencing sectors
(used for deep water and other tricks), these won't be tesselated,
causing HOM.

Explicitly detect self-referencing sectors and what sector
contains them (i.e. what should be rendered in their place),
and account for them during subsector tessellation.