Make GL fuzz effect emulate streakiness of software
The "streak" effect of software fuzz is due to runs of negative offsets
in the fuzz offset table causing the same pixel to be copied and
darkened several times in a row.  This shader emulates that using a
lookup table derived from the original offset table, as well as
darkness steps based on dcolors.c

This also changes the "time" uniform to a "count" uniform which is
incremented on each fuzzy sprite drawn, to ensure sprites drawn on
the same frame have different patterns.  The unused screen resolution
uniform was removed.

Things to make this even more accurate:

- Make the start offset of each column be the sum of non-transparent
  pixels in all previous columns, rather than random as it is now.
- Actually copy offset background pixels and go through colormaps
  to perform darkening.  This would require a major change to the
  rendering pipeline to use intermediate textures so previous pixels
  can be read.