Remove SMALLDELTA
It caused walls to be rendered a fraction of a unit off. 
This had already been done in 948c28a,  but was reverted in 8ed8316.  I tried all sorts of maps,  but couldnt replicate the mentioned "dots on seams between flats and walls",  so I assume that its a bug of the past.