Fix potential hang after restoring keyframe due to invalid blocklist pointers.
When restoring a keyframe, the blockmap is not cleared and blocklist pointers are modified during P_UnArchiveThinkers. The blockmap is then restored in UnArchiveBlockLinks, but this can leave blocklist pointers in an invalid state.

This can then lead to a hang later when the blocklist is iterated and may get stuck in an infinite loop.