Fix undefined behaviour
The MAPINFO parser has a couple of places where data is put into an
std::vector, and then a pointer to the first element of the vector
is taken through the "&vector[0]" pattern. Unfortunately, using
operator[] to access a non-existent element of std::vector is
undefined behaviour; which means that, should the vector be empty,
then nasal demons galore.

This commit adds "is the vector empty?" checks in relevant places.