CMake: Switch around logic to apply flags/defines.
Move from providing functions returning lists of flags and defines
which then needed to be applied to each target, to providing functions
applying the flags and defines to the target directly.

As a customisation point, some cache variables may be set if the default
does not work on a user's machine.

This is more in line with how users typically interact with CMake
targets, and avoid common typo issues such as:

```
get_flags(MY_COMPILE_FLAGS)
...
target_compile_options(my_target PRIVATE ${MY_COPILE_FLAGS})
```